+++ packages/sources/react/src/components/navigation/VtmnSearch/index.ts
@@ -1,0 +1,2 @@
+export { default as VtmnSearch } from './VtmnSearch';
+export * from './VtmnSearch';