+++ packages/sources/css/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.60.5](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css@0.60.4...@vtmn/css@0.60.5) (2022-03-18)
+
+**Note:** Version bump only for package @vtmn/css
+
+
+
+
+
 ## [0.60.4](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css@0.60.3...@vtmn/css@0.60.4) (2022-03-11)
 
 **Note:** Version bump only for package @vtmn/css